# perf(turbo-tasks): Mark a bunch of stuff as local