--- a/turbopack/crates/turbopack-ecmascript/src/references/amd.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/amd.rs
@@ -156,7 +156,7 @@ impl AmdDefineWithDependenciesCodeGen {
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for AmdDefineWithDependenciesCodeGen {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         module_graph: Vc<ModuleGraph>,