--- a/turbopack/crates/turbopack-ecmascript/src/references/esm/dynamic.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/esm/dynamic.rs
@@ -110,7 +110,7 @@ impl ChunkableModuleReference for EsmAsyncAssetReference {
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for EsmAsyncAssetReference {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         module_graph: Vc<ModuleGraph>,