--- a/turbopack/crates/turbopack-ecmascript/src/references/esm/meta.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/esm/meta.rs
@@ -97,7 +97,7 @@ impl ImportMetaRef {
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for ImportMetaRef {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         _module_graph: Vc<ModuleGraph>,