--- a/turbopack/crates/turbopack-ecmascript/src/references/esm/module_id.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/esm/module_id.rs
@@ -63,7 +63,7 @@ impl ChunkableModuleReference for EsmModuleIdAssetReference {
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for EsmModuleIdAssetReference {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         module_graph: Vc<ModuleGraph>,