--- a/turbopack/crates/turbopack-ecmascript/src/references/esm/module_item.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/esm/module_item.rs
@@ -35,7 +35,7 @@ impl EsmModuleItem {
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for EsmModuleItem {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         _module_graph: Vc<ModuleGraph>,