--- a/turbopack/crates/turbopack-ecmascript/src/references/ident.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/ident.rs
@@ -29,7 +29,7 @@ impl IdentReplacement {
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for IdentReplacement {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         _module_graph: Vc<ModuleGraph>,