--- a/turbopack/crates/turbopack-ecmascript/src/references/require_context.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/require_context.rs
@@ -288,7 +288,7 @@ impl ChunkableModuleReference for RequireContextAssetReference {}
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for RequireContextAssetReference {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         module_graph: Vc<ModuleGraph>,