--- a/turbopack/crates/turbopack-ecmascript/src/references/worker.rs
+++ b/turbopack/crates/turbopack-ecmascript/src/references/worker.rs
@@ -117,7 +117,7 @@ impl ChunkableModuleReference for WorkerAssetReference {}
 
 #[turbo_tasks::value_impl]
 impl CodeGenerateable for WorkerAssetReference {
-    #[turbo_tasks::function]
+    #[turbo_tasks::function(local)]
     async fn code_generation(
         &self,
         _module_graph: Vc<ModuleGraph>,