Cache hit numbers on dev + build: https://gist.github.com/bgw/1924f240a482377def4504e8d1fc3e18

Try marking a bunch of functions as local, see that no tests break. Measure numbers before/after:

**Mean memory consumption before (on canary):** 5.53 GiB
**Mean "compiled in" time before (on canary):** 18.7s

**Mean memory consumption after (this PR):** 5.67 GiB
**Mean "compiled in" time after (this PR):** 19.0s

<details>
Command run:

```
rm -rf .next && TURBOPACK=1 TURBOPACK_BUILD=1 TURBO_ENGINE_READ_ONLY=1 /bin/time -v pnpm next build --experimental-build-mode=compile
```

# Before, run 1

Compiled in: 19.1s

```
        Command being timed: "pnpm next build --experimental-build-mode=compile"
        User time (seconds): 121.55
        System time (seconds): 17.03
        Percent of CPU this job got: 513%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:27.00
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 5561800
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 61
        Minor (reclaiming a frame) page faults: 1090182
        Voluntary context switches: 1299111
        Involuntary context switches: 351305
        Swaps: 0
        File system inputs: 112584
        File system outputs: 737120
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

# Before, run 2

Compiled in: 18.9s

```
        Command being timed: "pnpm next build --experimental-build-mode=compile"
        User time (seconds): 118.63
        System time (seconds): 16.68
        Percent of CPU this job got: 503%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:26.88
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 6023472
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 2
        Minor (reclaiming a frame) page faults: 1420969
        Voluntary context switches: 1285934
        Involuntary context switches: 313524
        Swaps: 0
        File system inputs: 0
        File system outputs: 737168
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

# Before, run 3

Compiled in: 18.2s

```
        Command being timed: "pnpm next build --experimental-build-mode=compile"
        User time (seconds): 117.55
        System time (seconds): 14.76
        Percent of CPU this job got: 509%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:25.99
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 5835060
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 1
        Minor (reclaiming a frame) page faults: 1416925
        Voluntary context switches: 1229953
        Involuntary context switches: 341983
        Swaps: 0
        File system inputs: 0
        File system outputs: 737216
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

# After, run 1

Compiled in: 19.6s

```
        Command being timed: "pnpm next build --experimental-build-mode=compile"
        User time (seconds): 125.92
        System time (seconds): 17.09
        Percent of CPU this job got: 516%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:27.69
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 5853124
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 61
        Minor (reclaiming a frame) page faults: 1266331
        Voluntary context switches: 1289601
        Involuntary context switches: 361896
        Swaps: 0
        File system inputs: 101312
        File system outputs: 737456
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

# After, run 2

Compiled in: 18.9s

```
        Command being timed: "pnpm next build --experimental-build-mode=compile"
        User time (seconds): 123.56
        System time (seconds): 16.48
        Percent of CPU this job got: 519%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:26.95
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 5817016
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 1060747
        Voluntary context switches: 1309994
        Involuntary context switches: 363910
        Swaps: 0
        File system inputs: 0
        File system outputs: 736960
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```


# After, run 3

Compiled in: 18.5s

```
        Command being timed: "pnpm next build --experimental-build-mode=compile"
        User time (seconds): 119.21
        System time (seconds): 15.30
        Percent of CPU this job got: 510%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:26.36
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 6178220
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 1143656
        Voluntary context switches: 1297677
        Involuntary context switches: 363754
        Swaps: 0
        File system inputs: 0
        File system outputs: 737848
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```
</details>