@ijjk 2025-01-24
## Failing test suites <!--33461433548-->
Commit: 28324c5c9e578098c21e01a0289e8c49f455fa1d

`pnpm test-dev-turbo test/development/app-dir/hmr-asset-prefix-full-url/asset-prefix.test.ts` (turbopack)
- app-dir assetPrefix full URL > should not break HMR when asset prefix set to full URL

<details>
<summary>Expand output</summary>

  ● app-dir assetPrefix full URL › should not break HMR when asset prefix set to full URL

    expect(received).toBe(expected) // Object.is equality

    Expected: "after edit"
    Received: "before edit"

      27 |
      28 |     await retry(async () => {
    > 29 |       expect(await browser.elementByCss('p').text()).toBe('after edit')
         |                                                      ^
      30 |     })
      31 |   })
      32 | })

      at toBe (development/app-dir/hmr-asset-prefix-full-url/asset-prefix.test.ts:29:54)
      at retry (lib/next-test-utils.ts:810:14)
      at Object.<anonymous> (development/app-dir/hmr-asset-prefix-full-url/asset-prefix.test.ts:28:5)

</details>


Read more about building and testing Next.js in [contributing.md](https://github.com/vercel/next.js/blob/canary/contributing.md).